refactor: move Header into MainPage && isolating SearchBar
