+++ src/components/Header.js
@@ -5,7 +5,6 @@ import {
   colors,
   Grid,
   IconButton,
-  InputBase,
   List,
   ListItem,
   SwipeableDrawer,
@@ -15,12 +14,12 @@ import {
   Typography,
   withStyles,
 } from '@material-ui/core';
-import SearchIcon from '@material-ui/icons/Search';
 import MenuIcon from '@material-ui/icons/Menu';
-import Cancel from '@material-ui/icons/Cancel';
 import { fade } from '@material-ui/core/styles';
 
 import Menu from './Menu';
+import SearchBar from './SearchBar';
+import { total } from '../data';
 
 const logo = require('../assets/pron.png');
 
@@ -120,61 +119,6 @@ const styles = theme => ({
     minHeight: '4em',
     color: theme.palette.grey['600'],
   },
-  search: {
-    position: 'relative',
-    borderRadius: theme.shape.borderRadius,
-    backgroundColor: fade(theme.palette.common.black, 0.05),
-    '&:hover': {
-      backgroundColor: fade(theme.palette.common.black, 0.1),
-    },
-    width: '100%',
-    marginLeft: theme.spacing(3),
-    [theme.breakpoints.up('sm')]: {
-      width: 'auto',
-    },
-    [theme.breakpoints.down('xs')]: {
-      margin: theme.spacing(0, 2),
-    },
-  },
-  searchIcon: {
-    width: theme.spacing(7),
-    height: '100%',
-    position: 'absolute',
-    pointerEvents: 'none',
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  resetIcon: {
-    width: theme.spacing(4),
-    padding: theme.spacing(0, 0.5),
-    height: '100%',
-    position: 'absolute',
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'center',
-    right: 0,
-    top: 0,
-    opacity: 0.2,
-    '&:hover': {
-      opacity: 0.4,
-      cursor: 'pointer',
-    },
-  },
-  inputRoot: {
-    color: 'inherit',
-  },
-  inputInput: {
-    padding: theme.spacing(1, 1, 1, 6),
-    transition: theme.transitions.create('width'),
-    width: '100%',
-    [theme.breakpoints.up('sm')]: {
-      width: 180,
-      '&:focus': {
-        width: 240,
-      },
-    },
-  },
 });
 
 const MenuLink = (props, ref) => (
@@ -196,7 +140,6 @@ const MenuLink = (props, ref) => (
 class Header extends Component {
   state = {
     menuDrawer: false,
-    searchText: '',
   };
 
   componentDidMount() {
@@ -213,7 +156,6 @@ class Header extends Component {
 
   render() {
     const { classes } = this.props;
-
     return (
       <AppBar position="absolute" color="default" className={classes.appBar}>
         <Toolbar>
@@ -234,102 +176,70 @@ class Header extends Component {
                   </span>
                 </a>
               </Typography>
-              {!this.props.noTabs && (
-                <React.Fragment>
-                  <div className={classes.org}>
-                    <Typography
-                      variant="h6"
-                      noWrap
-                      style={{ textTransform: 'uppercase' }}
-                    >
-                      <a href={Menu[0].url} className={classes.link} target="_about">
-                        <img className={classes.logoImg} src={Menu[0].img} alt="campfire" />
-                        <span
-                          className={clsx(classes.appName, classes.campLabel)}
-                        >
-                          Campfire
-                        </span>
-                      </a>
-                    </Typography>
-                  </div>
-                  <div className={classes.search}>
-                    <div className={classes.searchIcon}>
-                      <SearchIcon />
-                    </div>
-                    <InputBase
-                      placeholder="Search..."
-                      classes={{
-                        root: classes.inputRoot,
-                        input: classes.inputInput,
-                      }}
-                      inputProps={{ 'aria-label': 'Search' }}
-                      onChange={(event) => {
-                        this.setState({ searchText: event.target.value });
-                        this.props.onSearch(event);
-                      }}
-                      value={this.state.searchText}
-                    />
-                    {this.state.searchText
-                      && (
-                        <span
-                          onMouseDown={() => {
-                            this.props.onReset(); this.setState({ searchText: '' });
-                          }}
-                          className={classes.resetIcon}
-                        >
-                          <Cancel />
-                        </span>
-                      )}
-                  </div>
-                  <div className={classes.iconContainer}>
-                    <IconButton onClick={this.mobileMenuOpen} className={classes.iconButton} color="inherit" aria-label="Menu">
-                      <MenuIcon />
-                    </IconButton>
-                  </div>
-                  <div className={classes.tabContainer}>
-                    <SwipeableDrawer anchor="right" open={this.state.menuDrawer} onClose={this.mobileMenuClose} onOpen={this.mobileMenuOpen}>
-                      <List>
-                        {Menu.map((item, index) => (
-                          <ListItem
-                            component={React.forwardRef(MenuLink)}
-                            url={item.url}
-                            name={item.name}
-                            target="_blank"
-                            img={item.img}
-                            classes={{ root: classes.listItem }}
-                            button
-                            key={index}
-                          />
-                        ))}
-                      </List>
-                    </SwipeableDrawer>
-                    <Tabs
-                      value={false}
-                      indicatorColor="primary"
-                      textColor="primary"
-                      onChange={this.handleChange}
-                      className={classes.itemsLayout}
-                    >
-                      {Menu.map((item, index) => {
-                        if (index !== 0) {
-                          return (
-                            <Tab
-                              component={React.forwardRef(MenuLink)}
-                              url={item.url}
-                              name={item.name}
-                              target="_blank"
-                              img={item.img}
-                              classes={{ root: classes.tabItem }}
-                              key={index}
-                            />
-                          );
-                        }
-                        return null;
-                      })}
-                    </Tabs>
-                  </div>
-                </React.Fragment>
-              )}
+              <div className={classes.org}>
+                <Typography
+                  variant="h6"
+                  noWrap
+                  style={{ textTransform: 'uppercase' }}
+                >
+                  <a href={Menu[0].url} className={classes.link} target="_about">
+                    <img className={classes.logoImg} src={Menu[0].img} alt="campfire" />
+                    <span className={clsx(classes.appName, classes.campLabel)}>Campfire</span>
+                  </a>
+                </Typography>
+              </div>
+              <SearchBar
+                placeholder={`Search in ${total} words...`}
+                onSearch={this.props.onSearch}
+                onReset={this.props.onReset}
+              />
+              <div className={classes.iconContainer}>
+                <IconButton onClick={this.mobileMenuOpen} className={classes.iconButton} color="inherit" aria-label="Menu">
+                  <MenuIcon />
+                </IconButton>
+              </div>
+              <div className={classes.tabContainer}>
+                <SwipeableDrawer anchor="right" open={this.state.menuDrawer} onClose={this.mobileMenuClose} onOpen={this.mobileMenuOpen}>
+                  <List>
+                    {Menu.map((item, index) => (
+                      <ListItem
+                        component={React.forwardRef(MenuLink)}
+                        url={item.url}
+                        name={item.name}
+                        target="_blank"
+                        img={item.img}
+                        classes={{ root: classes.listItem }}
+                        button
+                        key={index}
+                      />
+                    ))}
+                  </List>
+                </SwipeableDrawer>
+                <Tabs
+                  value={false}
+                  indicatorColor="primary"
+                  textColor="primary"
+                  onChange={this.handleChange}
+                  className={classes.itemsLayout}
+                >
+                  {Menu.map((item, index) => {
+                    if (index !== 0) {
+                      return (
+                        <Tab
+                          component={React.forwardRef(MenuLink)}
+                          url={item.url}
+                          name={item.name}
+                          target="_blank"
+                          img={item.img}
+                          classes={{ root: classes.tabItem }}
+                          key={index}
+                        />
+                      );
+                    }
+                    return null;
+                  })}
+                </Tabs>
+              </div>
             </Grid>
           </Grid>
         </Toolbar>
